Added a company search function
I implemented Google Place API in the search function of the application to search businesses all around the world. After selecting a business, the program sends a fetch request to Apify, so that an actor sifts through all the reviews and other relevant information of the company and puts it all into a JSON file, which is then returned to the client.